Add stylesheet.css and style header
In this commit, I created the style sheet for the landing page.

Ina addition to this, I added the style for the first section class:
Header.
This style rearranges the contents of the header to push all the links
to the right, while the header logo remains on the right.

in addition to this, The links where styled to remove the underline
decoration, and their spacing was also adjusted to meet requirements.