Add CSS to "first" section
In this commit I have coded in Css to possition the elements of the
first section in their places.

I added margin between the image and the text on the left, as well as
positioning it centrally with justify content.

In addition to this i added margin to the links on the right and the
logo on the left, to line it up with the text and image below it.

Lastly, I added padding and background color to the container "first",
more specifically padding to the top and bottom of the container.