Complete Information section
In this commit I have completed the information section.

The main heading with the class ".top", was adjusted to have the
font-size 36px, the boldness 900px, and the text aligned centrally.

The texts below all the images have also been altered to allow them
to fit directly under the image. In addition to this, the image divs
in their own right have been limited to a width of 200px, to allow
the text to wrap beneath it.

Lastly, I added a margin of 30px to the top of the entire section,
and a margin of 40px to the bottom of the section.