Add style to quote, call to action and footer.
In this commit, I added to style code to my quote section.
This included adding the required background color, the required font
and font size, as well as positioning th text in the center.

In addition to this, I styled the call to action and footer, by adding
the required background color and the positioning the text accordingly.

I also made some minor adjustments to the first section of the website.
I added minor margin space to between the header and subset 1, for a
more uniform look.